Force creation of empty candles on new day. refs #106
If we reach a new day without having a previous fetch trigger the
creation of a new day there is a possibility we reach a weird edge
case where the expected most recent candle has a higher candle start
point than the first candle in the batch. To prevent an infinite loop
we force empty candles to be created for the new day.